refactor: split recording.ts into UI rendering + IO/API (coherence 0.44)
Extract Chrome runtime messaging, mic permission flow, and storage
operations into recording-io.ts. UI rendering (showRecording, showIdle,
timer, approval state) stays in recording.ts.
